flask_app/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
Flask-Login==0.6.3
email-validator==2.2.0
Werkzeug==3.0.3
mysqlclient==2.2.4
PyMySQL==1.1.1
